+++ crates/ffi/src/lib.rs
@@ -201,6 +201,10 @@ impl FilterListManager {
             Ok(())
         })
     }
+
+    pub fn get_rules_count(&self, ids: Vec<FilterId>) -> AGResult<Vec<RulesCountByFilter>> {
+        self.wrap(move |flm| flm.get_rules_count(ids))
+    }
 }
 
 impl FilterListManager {