+++ crates/ffi/src/platforms/windows/AdGuard.FilterListManagerProtobuf/AdGuard.FilterListManagerProtobuf/ProtobufGeneratedImpl/FlmInterface.cs
@@ -108,5 +108,9 @@ namespace FilterListManager
     public partial class GetDisabledRulesResponse : IAGOuterError 
     {
     }
+
+    public partial class GetRulesCountResponse : IAGOuterError
+    {
+    }
 }