+++ crates/filter-list-manager/src/manager/update_filters_action.rs
@@ -418,6 +418,8 @@ mod tests {
         let fourth_filter_id = -100012;
         let fifth_filter_id = 5;
 
+        let user_rules_count_new = 2;
+
         let mut filter1 = FilterEntity::default();
         filter1.filter_id = Some(first_filter_id);
         filter1.group_id = 1;
@@ -427,6 +429,7 @@ mod tests {
             filter_id: first_filter_id,
             text: "Filter1\nNonFilter1".to_string(),
             disabled_text: "NonFilter1".to_string(),
+            rules_count: 0,
         });
         filter1.download_url = download_url1.to_string();
 
@@ -438,6 +441,7 @@ mod tests {
             filter_id: second_filter_id,
             text: "Filter2\nNonFilter2".to_string(),
             disabled_text: "NonFilter2".to_string(),
+            rules_count: 0,
         });
         filter2.download_url = download_url2.to_string();
 
@@ -449,6 +453,7 @@ mod tests {
             filter_id: third_filter_id,
             text: "Filter3\nNonFilter3".to_string(),
             disabled_text: "NonFilter3".to_string(),
+            rules_count: 0,
         });
         filter3.download_url = download_url3.to_string();
 
@@ -463,6 +468,7 @@ mod tests {
             filter_id: fourth_filter_id,
             text: "Filter4\nNonFilter4".to_string(),
             disabled_text: "NonFilter4".to_string(),
+            rules_count: 0,
         });
         filter4.download_url = download_url4.to_string();
 
@@ -481,6 +487,7 @@ mod tests {
             // Disabled rules weren't written that way.
             // It will be mistaken to put here non-empty string
             disabled_text: String::new(),
+            rules_count: user_rules_count_new,
         });
         filter5.download_url = download_url5.to_string();
 
@@ -532,24 +539,28 @@ mod tests {
             filter_id: first_filter_id,
             text: "Filter1_new\nNonFilter1_new".to_string(),
             disabled_text: rules1.disabled_text,
+            rules_count: user_rules_count_new,
         });
 
         let _ = write_rules(RulesListEntity {
             filter_id: second_filter_id,
             text: "Filter2_new\nNonFilter2_new".to_string(),
             disabled_text: rules2.clone().disabled_text,
+            rules_count: user_rules_count_new,
         });
 
         let (_, new_rules3) = write_rules(RulesListEntity {
             filter_id: third_filter_id,
             text: "Filter3_new\nNonFilter3_new".to_string(),
             disabled_text: rules3.disabled_text,
+            rules_count: user_rules_count_new,
         });
 
         let _ = write_rules(RulesListEntity {
             filter_id: fourth_filter_id,
             text: "Filter4_new\nNonFilter4_new".to_string(),
             disabled_text: rules4.clone().disabled_text,
+            rules_count: user_rules_count_new,
         });
 
         let mut conf = Configuration::default();
@@ -670,6 +681,8 @@ mod tests {
         let fourth_filter_id = -100012;
         let fifth_filter_id = 5;
 
+        let user_rules_count_new = 2;
+
         let mut filter1 = FilterEntity::default();
         filter1.filter_id = Some(first_filter_id);
         filter1.group_id = 1;
@@ -679,6 +692,7 @@ mod tests {
             filter_id: first_filter_id,
             text: "Filter1\nNonFilter1".to_string(),
             disabled_text: "NonFilter1".to_string(),
+            rules_count: 0,
         });
         filter1.download_url = download_url1.to_string();
 
@@ -690,6 +704,7 @@ mod tests {
             filter_id: second_filter_id,
             text: "Filter2\nNonFilter2".to_string(),
             disabled_text: "NonFilter2".to_string(),
+            rules_count: 0,
         });
         filter2.download_url = download_url2.to_string();
 
@@ -701,6 +716,7 @@ mod tests {
             filter_id: third_filter_id,
             text: "Filter3\nNonFilter3".to_string(),
             disabled_text: "NonFilter3".to_string(),
+            rules_count: 0,
         });
         filter3.download_url = download_url3.to_string();
 
@@ -715,6 +731,7 @@ mod tests {
             filter_id: fourth_filter_id,
             text: "Filter4\nNonFilter4".to_string(),
             disabled_text: "NonFilter4".to_string(),
+            rules_count: 0,
         });
         filter4.download_url = download_url4.to_string();
 
@@ -733,6 +750,7 @@ mod tests {
             // Disabled rules weren't written that way.
             // It will be mistaken to put here non-empty string
             disabled_text: String::new(),
+            rules_count: user_rules_count_new,
         });
         filter5.download_url = download_url5.to_string();
 
@@ -784,24 +802,28 @@ mod tests {
             filter_id: first_filter_id,
             text: "Filter1_new\nNonFilter1_new".to_string(),
             disabled_text: rules1.clone().disabled_text,
+            rules_count: user_rules_count_new,
         });
 
         let (_, new_rules2) = write_rules(RulesListEntity {
             filter_id: second_filter_id,
             text: "Filter2_new\nNonFilter2_new".to_string(),
             disabled_text: rules2.clone().disabled_text,
+            rules_count: user_rules_count_new,
         });
 
         let (_, new_rules3) = write_rules(RulesListEntity {
             filter_id: third_filter_id,
             text: "Filter3_new\nNonFilter3_new".to_string(),
             disabled_text: rules3.clone().disabled_text,
+            rules_count: user_rules_count_new,
         });
 
         let (_, new_rules4) = write_rules(RulesListEntity {
             filter_id: fourth_filter_id,
             text: "Filter4_new\nNonFilter4_new".to_string(),
             disabled_text: rules4.clone().disabled_text,
+            rules_count: user_rules_count_new,
         });
 
         let mut conf = Configuration::default();