+++ crates/filter-list-manager/src/storage/db_bootstrap.rs
@@ -1,11 +1,9 @@
 use crate::storage::constants::{
     CUSTOM_FILTERS_GROUP_ID, SERVICE_GROUP_ID, USER_RULES_FILTER_LIST_ID,
 };
-use crate::storage::entities::db_metadata_entity::DBMetadataEntity;
 use crate::storage::entities::filter_entity::FilterEntity;
 use crate::storage::entities::filter_group_entity::FilterGroupEntity;
 use crate::storage::entities::rules_list_entity::RulesListEntity;
-use crate::storage::repositories::db_metadata_repository::DBMetadataRepository;
 use crate::storage::repositories::filter_group_repository::FilterGroupRepository;
 use crate::storage::repositories::filter_repository::FilterRepository;
 use crate::storage::repositories::rules_list_repository::RulesListRepository;
@@ -40,6 +38,7 @@ fn create_user_rules_rules_list_entity() -> RulesListEntity {
         filter_id: USER_RULES_FILTER_LIST_ID,
         text: String::new(),
         disabled_text: String::new(),
+        rules_count: 0,
     }
 }
 
@@ -52,19 +51,11 @@ fn create_group_entity_for_custom_filters() -> FilterGroupEntity {
     }
 }
 
-/// Creates metadata info
-fn create_db_metadata() -> DBMetadataEntity {
-    DBMetadataEntity::default()
-}
-
 /// Makes specific queries to an empty database with a schema
 pub(crate) fn db_bootstrap(transaction: &mut Transaction) -> rusqlite::Result<()> {
     let filter_entity = make_user_rules_filter_entity();
     let rule_entity = create_user_rules_rules_list_entity();
     let custom_group = create_group_entity_for_custom_filters();
-    let metadata_entity = create_db_metadata();
-
-    DBMetadataRepository::save(transaction, &metadata_entity)?;
 
     FilterRepository::new().insert(transaction, &[filter_entity])?;
     RulesListRepository::new().insert(transaction, &[rule_entity])?;