+++ crates/filter-list-manager/src/storage/services/constants.rs
@@ -1,0 +1,5 @@
+/// This marks line as "non-rule" line: comment, directive, etc.
+pub const NON_RULE_MARKER: char = '!';
+
+/// Also, comment line can start from "# " sequence
+pub const EXTRA_COMMENT_MARKER: &str = "# ";