+++ platform/apple/AdGuardFLM/Sources/AdGuardFLMLib/GeneratedProto/flm_interface.pb.swift
@@ -329,6 +329,18 @@ public struct FilterListManager_SetProxyModeRequest: Sendable {
   public init() {}
 }
 
+public struct FilterListManager_GetRulesCountRequest: Sendable {
+  // SwiftProtobuf.Message conformance is added in an extension below. See the
+  // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
+  // methods supported on all messages.
+
+  public var ids: [Int32] = []
+
+  public var unknownFields = SwiftProtobuf.UnknownStorage()
+
+  public init() {}
+}
+
 public struct FilterListManager_EmptyRequest: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
@@ -877,6 +889,29 @@ public struct FilterListManager_GetDisabledRulesResponse: Sendable {
   // methods supported on all messages.
 
   public var rulesRaw: [FilterListManager_DisabledRulesRaw] = []
+
+  public var error: FilterListManager_AGOuterError {
+    get {return _error ?? FilterListManager_AGOuterError()}
+    set {_error = newValue}
+  }
+  /// Returns true if `error` has been explicitly set.
+  public var hasError: Bool {return self._error != nil}
+  /// Clears the value of `error`. Subsequent reads from it will return its default value.
+  public mutating func clearError() {self._error = nil}
+
+  public var unknownFields = SwiftProtobuf.UnknownStorage()
+
+  public init() {}
+
+  fileprivate var _error: FilterListManager_AGOuterError? = nil
+}
+
+public struct FilterListManager_GetRulesCountResponse: Sendable {
+  // SwiftProtobuf.Message conformance is added in an extension below. See the
+  // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
+  // methods supported on all messages.
+
+  public var rulesCountByFilter: [FilterListManager_RulesCountByFilter] = []
 
   public var error: FilterListManager_AGOuterError {
     get {return _error ?? FilterListManager_AGOuterError()}
@@ -1648,6 +1683,38 @@ extension FilterListManager_SetProxyModeRequest: SwiftProtobuf.Message, SwiftPro
   public static func ==(lhs: FilterListManager_SetProxyModeRequest, rhs: FilterListManager_SetProxyModeRequest) -> Bool {
     if lhs.mode != rhs.mode {return false}
     if lhs.customProxyAddr != rhs.customProxyAddr {return false}
+    if lhs.unknownFields != rhs.unknownFields {return false}
+    return true
+  }
+}
+
+extension FilterListManager_GetRulesCountRequest: SwiftProtobuf.Message, SwiftProtobuf._MessageImplementationBase, SwiftProtobuf._ProtoNameProviding {
+  public static let protoMessageName: String = _protobuf_package + ".GetRulesCountRequest"
+  public static let _protobuf_nameMap: SwiftProtobuf._NameMap = [
+    1: .same(proto: "ids"),
+  ]
+
+  public mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
+    while let fieldNumber = try decoder.nextFieldNumber() {
+      // The use of inline closures is to circumvent an issue where the compiler
+      // allocates stack space for every case branch when no optimizations are
+      // enabled. https://github.com/apple/swift-protobuf/issues/1034
+      switch fieldNumber {
+      case 1: try { try decoder.decodeRepeatedInt32Field(value: &self.ids) }()
+      default: break
+      }
+    }
+  }
+
+  public func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
+    if !self.ids.isEmpty {
+      try visitor.visitPackedInt32Field(value: self.ids, fieldNumber: 1)
+    }
+    try unknownFields.traverse(visitor: &visitor)
+  }
+
+  public static func ==(lhs: FilterListManager_GetRulesCountRequest, rhs: FilterListManager_GetRulesCountRequest) -> Bool {
+    if lhs.ids != rhs.ids {return false}
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -2548,6 +2615,48 @@ extension FilterListManager_GetDisabledRulesResponse: SwiftProtobuf.Message, Swi
 
   public static func ==(lhs: FilterListManager_GetDisabledRulesResponse, rhs: FilterListManager_GetDisabledRulesResponse) -> Bool {
     if lhs.rulesRaw != rhs.rulesRaw {return false}
+    if lhs._error != rhs._error {return false}
+    if lhs.unknownFields != rhs.unknownFields {return false}
+    return true
+  }
+}
+
+extension FilterListManager_GetRulesCountResponse: SwiftProtobuf.Message, SwiftProtobuf._MessageImplementationBase, SwiftProtobuf._ProtoNameProviding {
+  public static let protoMessageName: String = _protobuf_package + ".GetRulesCountResponse"
+  public static let _protobuf_nameMap: SwiftProtobuf._NameMap = [
+    1: .standard(proto: "rules_count_by_filter"),
+    2: .same(proto: "error"),
+  ]
+
+  public mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
+    while let fieldNumber = try decoder.nextFieldNumber() {
+      // The use of inline closures is to circumvent an issue where the compiler
+      // allocates stack space for every case branch when no optimizations are
+      // enabled. https://github.com/apple/swift-protobuf/issues/1034
+      switch fieldNumber {
+      case 1: try { try decoder.decodeRepeatedMessageField(value: &self.rulesCountByFilter) }()
+      case 2: try { try decoder.decodeSingularMessageField(value: &self._error) }()
+      default: break
+      }
+    }
+  }
+
+  public func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
+    // The use of inline closures is to circumvent an issue where the compiler
+    // allocates stack space for every if/case branch local when no optimizations
+    // are enabled. https://github.com/apple/swift-protobuf/issues/1034 and
+    // https://github.com/apple/swift-protobuf/issues/1182
+    if !self.rulesCountByFilter.isEmpty {
+      try visitor.visitRepeatedMessageField(value: self.rulesCountByFilter, fieldNumber: 1)
+    }
+    try { if let v = self._error {
+      try visitor.visitSingularMessageField(value: v, fieldNumber: 2)
+    } }()
+    try unknownFields.traverse(visitor: &visitor)
+  }
+
+  public static func ==(lhs: FilterListManager_GetRulesCountResponse, rhs: FilterListManager_GetRulesCountResponse) -> Bool {
+    if lhs.rulesCountByFilter != rhs.rulesCountByFilter {return false}
     if lhs._error != rhs._error {return false}
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true